Splits out wolf ai triggers into separate files
