+++ Danki2/Assets/Scripts/AI/Triggers/DistanceGreaterThan.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0f5eec00753dc7e4f8a5a8c5796badc9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
