+++ Danki2/Assets/Scripts/AI/Triggers/DistanceLessThan.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1530b447a07da10469d2e6668eab863b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
