+++ Danki2/Assets/Scripts/AI/Triggers/HealthGoesBelow.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3c6d6fb8d6093ed4d80948228b174bc5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
