+++ Danki2/Assets/Scripts/AI/Triggers/HearsHowl.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5ea678ba621471e489a30bb94c090f3f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
