+++ Danki2/Assets/Scripts/AI/Triggers/InstantTrigger.cs
@@ -1,0 +1,6 @@
+﻿public class InstantTrigger : IAiTrigger
+{
+    public void Activate() {}
+    public void Deactivate() {}
+    public bool Triggers() => true;
+}
