+++ Danki2/Assets/Scripts/AI/Triggers/InstantTrigger.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a6fae44e34aae164c9b7d93f88c6f111
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
