+++ Danki2/Assets/Scripts/AI/Triggers/RandomTimeElapsed.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 478ed7dc67ad4734fb6c079a00daf13b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
