+++ Danki2/Assets/Scripts/AI/Triggers/TakesDamage.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9369d0100eeb8cc48b819d9189e62706
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
