+++ Danki2/Assets/Scripts/AI/Triggers/TimeElapsed.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9b4ffda27f4d55942a1e8ce356a3252f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
