+++ Danki2/Assets/Scripts/AI/Triggers/Wolf.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 54a043bf93bfdff41807251b0a02f6a2
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
