+++ Danki2/Assets/Scripts/AI/Triggers/Wolf/WolfBiteFinished.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 62940c6972197e045aa38215e58a743a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
