+++ Danki2/Assets/Scripts/AI/Triggers/Wolf/WolfPounceFinished.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b8c8ac5970427d84e87a5475936aef4f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
