+++ Danki2/Assets/Scripts/AI/Triggers/Wolf/WolfRandomAttackCountReached.cs
@@ -1,0 +1,39 @@
+﻿using System.Collections.Generic;
+using UnityEngine;
+
+public class WolfRandomAttackCountReached : IAiTrigger
+{
+    private readonly Wolf wolf;
+    private readonly int minAttacks;
+    private readonly int maxAttacks;
+
+    private int attacks;
+    private int requiredAttacks;
+    private readonly List<Subscription> attackSubscriptions = new List<Subscription>();
+
+    public WolfRandomAttackCountReached(Wolf wolf, int minAttacks, int maxAttacks)
+    {
+        this.wolf = wolf;
+        this.minAttacks = minAttacks;
+        this.maxAttacks = maxAttacks;
+    }
+
+    public void Activate()
+    {
+        attacks = 0;
+        requiredAttacks = Random.Range(minAttacks, maxAttacks + 1);
+        attackSubscriptions.Add(wolf.OnBite.Subscribe(() => attacks++));
+        attackSubscriptions.Add(wolf.OnPounce.Subscribe(() => attacks++));
+    }
+
+    public void Deactivate()
+    {
+        attackSubscriptions.ForEach(s => s.Unsubscribe());
+        attackSubscriptions.Clear();
+    }
+
+    public bool Triggers()
+    {
+        return attacks >= requiredAttacks;
+    }
+}
