+++ Danki2/Assets/Scripts/AI/Triggers/Wolf/WolfRandomAttackCountReached.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8a4eaebe771f5a54cb8398927817d29d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
