--- Danki2/Assets/Scripts/AI/Triggers/WolfTriggers.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: ad4e38faf3994f2f83848fddc9505d8c
-timeCreated: 1603838659
